requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==18.2.0
cffi==1.11.5
jinja2==2.10
markupsafe==1.0           # via jinja2
pycparser==2.19           # via cffi
